Handle smart account detection for inApp and ecosystem wallets
