+++ .changeset/two-bugs-rescue.md
@@ -1,0 +1,5 @@
+---
+"thirdweb": patch
+---
+
+Handle smart account detection for inApp and ecosystem wallets
